requirements.txt
fuzzywuzzy
aiogram==3.0.0b6
aiogram-deta @ git+https://github.com/mamsdeveloper/deta.git@webapp_included
odetam @ git+https://github.com/mamsdeveloper/ODetaM.git@async